Replace CRA with minimal app
Create React App is great but doesn't give you access to linting and
babel settings, as well as containing many things I don't care about.
This minimal react hmr setup should do the trick instead.
