--- frontend/Dockerfile
@@ -1,8 +1,0 @@
-FROM node:13-alpine
-
-WORKDIR /app
-EXPOSE 3000
-CMD ["npm", "run", "start"]
-
-RUN npm ci
-COPY . .
