+++ frontend/package.json
@@ -1,41 +1,48 @@
 {
-  "name": "create-react-app",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@lingui/react": "^3.0.0-7",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-scripts": "3.0.1"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.1.2",
-    "@lingui/cli": "^3.0.0-3",
-    "@lingui/loader": "^3.0.0-3",
-    "@lingui/macro": "^3.0.0-3",
-    "babel-core": "^7.0.0-bridge.0"
-  },
+  "name": "webpack-test",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "build": "webpack --env.mode=production --config webpack.config.js",
+    "dev": "webpack-dev-server --hot --env.mode=development",
+    "start": "serve -l 3000 -s dist",
+    "test": "jest",
+    "add-locale": "lingui add-locale",
     "extract": "lingui extract",
     "compile": "lingui compile"
   },
-  "eslintConfig": {
-    "extends": "react-app"
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@babel/runtime": "^7.7.1",
+    "@lingui/react": "^3.0.0-7",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
+    "serve": "^11.2.0"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "devDependencies": {
+    "@babel/core": "^7.7.0",
+    "@babel/plugin-transform-runtime": "^7.6.2",
+    "@babel/preset-env": "^7.7.1",
+    "@babel/preset-react": "^7.7.0",
+    "@hot-loader/react-dom": "^16.10.2",
+    "@lingui/cli": "^3.0.0-7",
+    "@lingui/loader": "^3.0.0-7",
+    "@lingui/macro": "^3.0.0-7",
+    "@testing-library/react": "^9.3.2",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-macros": "^2.6.1",
+    "fs-extra": "^8.1.0",
+    "html-webpack-plugin": "^3.2.0",
+    "html-webpack-root-plugin": "^0.10.0",
+    "jest": "^24.9.0",
+    "react-hot-loader": "^4.12.15",
+    "webpack": "^4.41.2",
+    "webpack-cli": "^3.3.10",
+    "webpack-config-utils": "^2.3.1",
+    "webpack-dev-server": "^3.9.0"
   }
 }
