--- frontend/src/LocaleSwitcher.css
@@ -1,19 +1,0 @@
-.LocaleSwitcher {
-  padding: 0;
-  list-style: none;
-  text-align: center;
-}
-
-.LocaleSwitcher li {
-  display: inline;
-}
-
-.Locale {
-  cursor: pointer;
-
-  border: none;
-  background: transparent;
-  color: #61dafb;
-
-  font-size: 0.8em;
-}
