+++ frontend/src/LocaleSwitcher.js
@@ -1,7 +1,6 @@
 import * as React from "react"
 
 import { locales, activate } from "./i18n.config"
-import "./LocaleSwitcher.css"
 
 export function LocaleSwitcher() {
   return (
