+++ frontend/src/i18n.config.js
@@ -1,17 +1,27 @@
-import { i18n } from "@lingui/core"
+import { i18n } from '@lingui/core'
 
 export const locales = {
-  en: "English",
-  fr: "Français"
+  en: 'English',
+  fr: 'Français',
 }
 
 export async function activate(locale) {
-  const catalog = await import(
-    /* webpackChunkName: "i18n-[index]" */ `@lingui/loader!./locales/${locale}.po`
-  )
+  let catalog
+  if (process.env.NODE_ENV === 'development') {
+    catalog = await import(
+      /* webpackMode: "lazy", webpackChunkName: "i18n-[index]" */
+      `@lingui/loader!./locales/${locale}.po`
+    )
+  } else {
+    // for production or test use js:
+    catalog = await import(
+      /* webpackMode: "lazy", webpackChunkName: "i18n-[index]" */
+      `./locales/${locale}.js`
+    )
+  }
 
   i18n.load(locale, catalog)
   i18n.activate(locale)
 }
 
-activate("en")
+activate('en')
