+++ frontend/src/index.js
@@ -1,12 +1,8 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
+import { hot } from 'react-hot-loader/root'
+import React from 'react'
+import { render } from 'react-dom'
+import { App } from './App'
 
-ReactDOM.render(<App />, document.getElementById('root'));
+hot(<App />)
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+render(<App />, document.getElementById('root'))
