+++ frontend/src/locales/en.po
@@ -1,6 +1,6 @@
 msgid ""
 msgstr ""
-"POT-Creation-Date: 2018-10-13 13:51+0200\n"
+"POT-Creation-Date: 2019-10-27 19:00-0400\n"
 "Mime-Version: 1.0\n"
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
@@ -13,14 +13,6 @@ msgstr ""
 "Language-Team: \n"
 "Plural-Forms: \n"
 
-#: src/App.js:17
-msgid "Edit <0>src/App.js</0> and save to reload."
-msgstr "Edit <0>src/App.js</0> and save to reload."
-
-#: src/App.js:27
-msgid "Learn React"
-msgstr "Learn React"
-
-#: src/App.js:15
-msgid "logo"
-msgstr "logo"
+#: src/App.js:14
+msgid "Hello from React"
+msgstr "Hello from React"
