+++ frontend/webpack.config.js
