--- react-webpack/README.md
@@ -1,11 +1,0 @@
-# Minimum Viable React+Webpack
-
-There are many boilerplate projects. This one is mine. The goal here is simply
-fitting together the basic pieces for a Single Page Application: React, Babel and Webpack.
-
-Where this differs from other boilerplate projects is:
-
-* builds for production with `npm run build`
-* run a HMR dev-server on port 3000 with `npm run dev`
-* Has enough Babel stuff to cope with React.
-* It doesn't do anything else.
