--- react-webpack/lingui.config.js
@@ -1,11 +1,0 @@
-module.exports = {
-  locales: ["en", "fr"],
-  catalogs: [
-    {
-      path: "src/locales/{locale}",
-      include: ["src"]
-    }
-  ],
-  format: "po",
-  sourceLocale: "en"
-}
