--- react-webpack/src/locales/en.po
@@ -1,12 +1,0 @@
-msgid ""
-msgstr ""
-"POT-Creation-Date: 2019-10-27 19:00-0400\n"
-"Mime-Version: 1.0\n"
-"Content-Type: text/plain; charset=utf-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"X-Generator: @lingui/cli\n"
-"Language: en\n"
-
-#: src/App.js:14
-msgid "Hello from React"
-msgstr "Hello from React"
