--- react-webpack/src/locales/fr.js
@@ -1,1 +1,0 @@
-/* eslint-disable */module.exports={localeData:{"plurals":function(n,ord){if(ord)return n==1?"one":"other";return n>=0&&n<2?"one":"other"}},messages:{"Hello from React":"Bonjour de React"}};
